requirements.txt
Flask>=2.0
flask-cors
Pillow
gunicorn
pikepdf==8.1.0
flask-cors 
pymupdf
img2pdf   
pdf2docx
werkzeug>=2.0
docx2pdf
pywin32
python-pptx
PyPDF2
werkzeug
reportlab
pillow-heif
cairosvg
rembg
onnxruntime
moviepy==1.0.3
numpy
yt-dlp
PyPDF2
python-docx
